Consolidate workflow step validation and evaluation in core/workflow.
Move shared value-ref and when-clause validation into core/workflow, split step proto conversion into convert_workflow_steps.go, and extract the template/value evaluator from bootstrap runtime glue.
